+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-26 03:02:31
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 0053ca63ffda2df1af0c9baeb8407ec4107a17cc
+
+## [ForwardWidgets-2kuai] - 2025-08-26 03:02:31
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: c0d8253cf0b86f70922ed58507f435d37fc7d5ce
+
+## [ForwardWidgets-pack1r] - 2025-08-26 03:02:31
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: da7271af29fa649347ea4e176569aca0c35f1816
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-25 03:07:50
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main